Time: 82 ms (5.28%), Space: 93.1 MB (5.28%) - LeetHub

+++ 2425-bitwise-xor-of-all-pairings/2425-bitwise-xor-of-all-pairings.cpp
@@ -3,7 +3,7 @@ public:
     int xorAllNums(vector<int>& nums1, vector<int>& nums2) {
         int n = nums1.size();
         int m = nums2.size();
-        unordered_map<int,int> mp;
+        unordered_map<int,long long int > mp;
         for(auto num : nums1){
             mp[num] +=m;
         }
